Add GitHub Actions CI
Based on this old approach: [0]

Note, that the --global arg to install cake had to be added (despite not
mentioned here[1]) because otherwise dotnet would throw this error[2].

[0] https://github.com/nblockchain/DotNetEssentials/commit/cce59d5c1d08f757b8943a4f8aa9a10804381b88
[1] https://cakebuild.net/docs/getting-started/bootstrapping-scripts#bootstrapping-for.net-core-tool
[2]
Cannot find a manifest file.
For a list of locations searched, specify the "-d" option before the tool name.
If you intended to install a global tool, add `--global` to the command.
If you would like to create a manifest, use `dotnet new tool-manifest`, usually in the repo root directory.
Cannot find a manifest file.
For a list of locations searched, specify the "-d" option before the tool name.
No tools were restored.
Could not execute because the specified command or file was not found.
Possible reasons for this include:
  * You misspelled a built-in dotnet command.
  * You intended to execute a .NET Core program, but dotnet-cake does not exist.
  * You intended to run a global tool, but a dotnet-prefixed executable with this name could not be found on the PATH.